+++ mehari/src/db/create/mod.rs
@@ -5,38 +5,38 @@ use crate::common::trace_rss_now;
 use crate::pbs::txs::{SourceVersion, TxSeqDatabase};
 use anyhow::{Error, anyhow};
 use cli::Args;
-use derive_new::new;
 use enumflags2::{BitFlag, BitFlags};
 use hgvs::data::cdot::json::models as cdot_models;
-use hgvs::data::cdot::json::models::{BioType, Gene, GenomeAlignment, Tag, Transcript};
-use hgvs::sequences::{TranslationTable, translate_cds};
+use hgvs::data::cdot::json::models::{BioType, Gene, Tag, Transcript};
 use itertools::Itertools;
 use models::{GeneId, Identifier, LabelEntry, ReportEntry};
 use once_cell::sync::Lazy;
 use prost::Message;
-use rayon::iter::Either;
-use rayon::prelude::*;
-use reference::SequenceProvider;
-use seqrepo::AliasOrSeqId;
 use serde_json::json;
 use serde_with::DisplayFromStr;
 use std::cmp::{PartialEq, Reverse};
 use std::collections::{HashMap, HashSet};
 use std::fs::File;
 use std::io::BufWriter;
-use std::ops::Not;
 use std::path::Path;
 use std::{io::Write, time::Instant};
 use strum::Display;
 use thousands::Separable;
 
+mod build;
 mod cdot;
 pub mod cli;
+mod filter;
 mod gff3;
 pub mod models;
 pub mod reference;
 
+use crate::db::create::build::build_protobuf;
 use crate::db::create::cdot::load_cdot;
+use crate::db::create::filter::{
+    filter_empty_gene_id_mappings, filter_genes, filter_initial_gene_id_entries,
+    filter_transcripts, filter_transcripts_with_sequence,
+};
 use crate::db::create::gff3::load_gff3;
 use models::*;
 
@@ -178,41 +178,6 @@ impl TranscriptLoader {
         self
     }
 
-    fn filter_initial_gene_id_entries(&mut self) -> Result<(), Error> {
-        if self.disable_filters {
-            return Ok(());
-        }
-        tracing::info!("Filtering Gene Id entries …");
-        for (gene_id, tx_ids) in &self.gene_id_to_transcript_ids {
-            if tx_ids.is_empty() {
-                *self
-                    .discards
-                    .entry(Identifier::Gene(gene_id.clone()))
-                    .or_default() |= Reason::NoTranscripts;
-            }
-            if let Some(gene) = self.gene_id_to_gene.get(gene_id) {
-                if gene
-                    .biotype
-                    .as_ref()
-                    .is_some_and(|bt| bt.iter().any(|b| DISCARD_BIOTYPES_GENES.contains(b)))
-                {
-                    *self
-                        .discards
-                        .entry(Identifier::Gene(gene_id.clone()))
-                        .or_default() |= Reason::Biotype;
-                }
-            } else {
-                *self
-                    .discards
-                    .entry(Identifier::Gene(gene_id.clone()))
-                    .or_default() |= Reason::MissingGene;
-            }
-        }
-        // self.discard()?;
-        tracing::info!("… done filter Gene Id entries");
-        Ok(())
-    }
-
     fn apply_fixes(
         &mut self,
         transcript_id_to_tags: &Option<HashMap<TranscriptId, Vec<FeatureTag>>>,
@@ -222,482 +187,6 @@ impl TranscriptLoader {
             self.update_transcript_tags(txid_to_label);
         }
         self.fix_cds();
-    }
-
-    fn filter_genes(&mut self) -> Result<(), Error> {
-        if self.disable_filters {
-            return Ok(());
-        }
-        tracing::info!("Filtering genes …");
-        // Check whether the gene has a gene symbol.
-        let missing_symbol = |_gene_id: GeneId, gene: &Gene, _txs: &[&Transcript]| -> bool {
-            gene.gene_symbol.is_none() || gene.gene_symbol.as_ref().unwrap().is_empty()
-        };
-        // Check whether the gene has a biotype that should be discarded.
-        let biotype = |_gene_id: GeneId, gene: &Gene, _txs: &[&Transcript]| -> bool {
-            gene.biotype
-                .as_ref()
-                .map(|bt| bt.iter().any(|bt| DISCARD_BIOTYPES_GENES.contains(bt)))
-                .unwrap_or(false)
-        };
-        // Check whether all transcripts for a gene are predicted.
-        let predicted_only = |_gene_id: GeneId, _gene: &Gene, txs: &[&Transcript]| -> bool {
-            !txs.is_empty() && txs.iter().all(|tx| tx.id.starts_with('X'))
-        };
-        type Filter = fn(GeneId, &Gene, &[&Transcript]) -> bool;
-        let filters: [(Filter, Reason); 3] = [
-            (missing_symbol, Reason::MissingGeneSymbol),
-            (biotype, Reason::Biotype),
-            (predicted_only, Reason::PredictedTranscriptsOnly),
-        ];
-
-        let discarded_genes = self
-            .gene_id_to_gene
-            .iter()
-            .filter_map(|(gene_id, gene)| {
-                let txs = self
-                    .gene_id_to_transcript_ids
-                    .get(gene_id)
-                    .map(|tx_ids| {
-                        tx_ids
-                            .iter()
-                            .filter_map(|tx_id| self.transcript_id_to_transcript.get(tx_id))
-                            .collect_vec()
-                    })
-                    .unwrap_or_default();
-                let reason = filters
-                    .iter()
-                    .filter_map(|(f, r)| f(gene_id.clone(), gene, &txs).then_some(*r))
-                    .fold(BitFlags::<Reason>::default(), |a, b| a | b);
-                let empty = reason.is_empty();
-                (!empty).then_some((Identifier::Gene(gene_id.clone()), reason))
-            })
-            .collect_vec();
-
-        for (id, reason) in discarded_genes {
-            self.mark_discarded(&id, reason)?;
-        }
-        // self.discard()?;
-        tracing::info!("… done filtering genes");
-        Ok(())
-    }
-
-    /// Filter transcripts for gene.
-    ///
-    /// We employ the following rules:
-    ///
-    /// - Remove redundant transcripts with the same identifier and pick only the
-    ///   transcripts that have the highest version number for one assembly.
-    /// - Do not pick any `XM_`/`XR_` (NCBI predicted only) transcripts.
-    /// - Do not pick any `NR_` transcripts when there are coding `NM_` transcripts.
-    fn filter_transcripts(&mut self) -> Result<(), Error> {
-        if self.disable_filters {
-            return Ok(());
-        }
-        tracing::info!("Filtering transcripts …");
-
-        /// Params used as args for filter functions defined below
-        #[derive(new)]
-        struct Params<'a> {
-            tx: &'a Transcript,
-            tx_id: &'a str,
-        }
-
-        // Check whether the transcript has an HGNC id.
-        let missing_hgnc = |p: &Params| -> bool {
-            p.tx.hgnc.is_none()
-                || p.tx.hgnc.as_ref().unwrap().is_empty()
-                || p.tx.hgnc.as_ref().unwrap().starts_with("GENE:")
-        };
-
-        // Check whether the transcript has any genome builds.
-        let empty_genome_builds = |p: &Params| -> bool { p.tx.genome_builds.is_empty() };
-
-        // Check whether the transcript has tags cds_start_NF or cds_end_NF.
-        let cds_start_end_nf = |p: &Params| -> bool {
-            p.tx.genome_builds.values().any(|gb| {
-                if let Some(tag) = &gb.tag {
-                    tag.contains(&Tag::Other("cds_start_NF".into()))
-                        || tag.contains(&Tag::Other("cds_end_NF".into()))
-                } else {
-                    false
-                }
-            })
-        };
-
-        // Check whether the transcript is marked as partial.
-        // Ignore "partial" tag for NR transcripts.
-        // For protein coding transcripts which are marked as partial,
-        // check whether there's at least 1bp of 5' and 3' UTR
-        // (and then discard the 'partial' tag).
-        let partial = |p: &Params| -> bool {
-            if !matches!(p.tx.partial, Some(1)) {
-                return false;
-            }
-            if p.tx_id.starts_with("NR") {
-                false
-            } else if p.tx.protein_coding() {
-                assert_eq!(p.tx.genome_builds.len(), 1);
-                let alignment = p.tx.genome_builds.values().next().unwrap();
-                let exons = alignment.exons.clone();
-                if exons.is_empty() {
-                    return true;
-                }
-                let five_prime_trunc = {
-                    let cds_start = alignment.cds_start;
-                    let first = exons.iter().min_by_key(|e| e.alt_start_i).unwrap();
-                    Some(first.alt_start_i) == cds_start
-                };
-                let three_prime_trunc = {
-                    let cds_end = alignment.cds_end;
-                    let last = exons.iter().max_by_key(|e| e.alt_end_i).unwrap();
-                    Some(last.alt_end_i) == cds_end
-                };
-                five_prime_trunc && three_prime_trunc
-            } else {
-                true
-            }
-        };
-
-        // Check whether the transcript is predicted.
-        let predicted = |p: &Params| -> bool { p.tx_id.starts_with('X') };
-
-        // Check whether the transcript has a biotype that should be discarded.
-        let biotype = |p: &Params| -> bool {
-            p.tx.biotype
-                .as_ref()
-                .map(|bt| {
-                    bt.iter()
-                        .any(|bt| DISCARD_BIOTYPES_TRANSCRIPTS.contains(bt))
-                })
-                .unwrap_or(false)
-        };
-
-        // We have two sets of filters here:
-        // `tx_filters`: for every transcript known to us at this point (so every transcript from cdot)
-        // `gene_id_grouped_tx_filters`: all transcripts associated to a gene id
-        // The latter set relies on the grouping of transcripts by gene id,
-        // e.g. to discard transcripts with an older version within the same gene id group
-        type Filter = fn(&Params) -> bool;
-        let tx_filters: [(Filter, Reason); 6] = [
-            (missing_hgnc, Reason::MissingGeneId),
-            (empty_genome_builds, Reason::EmptyGenomeBuilds),
-            (partial, Reason::OnlyPartialAlignmentInRefSeq),
-            (predicted, Reason::PredictedTranscript),
-            (biotype, Reason::Biotype),
-            (cds_start_end_nf, Reason::CdsStartOrEndNotConfirmed),
-        ];
-
-        #[derive(new)]
-        struct GroupParams<'a> {
-            idx: usize,
-            tx_ids: &'a [TranscriptId],
-            txs: &'a [&'a Transcript],
-        }
-        type GroupFilter = fn(&GroupParams) -> bool;
-
-        // Check whether we have a newer version of the same transcript accession available.
-        let old_version = |p: &GroupParams| -> bool {
-            let versioned = Self::by_release_and_version(p.txs);
-            let (tx_ac, tx_version) = p.tx_ids[p.idx].split_version();
-            for release in p.txs[p.idx].genome_builds.keys() {
-                if let Some(other) = versioned.get(&(release.into(), tx_ac.to_string()))
-                    && other.iter().any(|(version, _tx)| *version > tx_version)
-                {
-                    return true;
-                }
-            }
-            false
-        };
-
-        // Check whether we have already seen an NM transcript for the gene.
-        let nm_instead_of_nr = |p: &GroupParams| -> bool {
-            p.tx_ids[p.idx].starts_with("NR_")
-                && p.tx_ids.iter().any(|tx_id| tx_id.starts_with("NM_"))
-        };
-
-        let gene_id_grouped_tx_filters: [(GroupFilter, Reason); 2] = [
-            (old_version, Reason::OldVersion),
-            (nm_instead_of_nr, Reason::UseNmTranscriptInsteadOfNr),
-        ];
-
-        let start = Instant::now();
-
-        // Apply first set of filters (which do not depend on gene id grouping)
-        let discarded = self
-            .transcript_id_to_transcript
-            .iter()
-            .filter_map(|(tx_id, tx)| {
-                let p = Params::new(tx, tx_id);
-                let reason = tx_filters
-                    .iter()
-                    .filter_map(|(f, r)| f(&p).then_some(*r))
-                    .fold(BitFlags::<Reason>::default(), |a, b| a | b);
-                let empty = reason.is_empty();
-                (!empty).then(|| (Identifier::Transcript(tx_id.clone()), reason))
-            })
-            .collect_vec();
-
-        for (id, reason) in discarded {
-            self.mark_discarded(&id, reason)?;
-        }
-
-        // Apply second set of filters (which depend on gene id grouping)
-        let discarded = self
-            .gene_id_to_transcript_ids
-            .iter()
-            .filter_map(|(_gene_id, tx_ids)| {
-                let txs = tx_ids
-                    .iter()
-                    .map(|tx_id| self.transcript_id_to_transcript.get(tx_id).unwrap())
-                    .collect_vec();
-                let tx_reasons = tx_ids
-                    .iter()
-                    .enumerate()
-                    .filter_map(|(idx, tx)| {
-                        let p = GroupParams::new(idx, tx_ids, &txs);
-                        let reason = gene_id_grouped_tx_filters
-                            .iter()
-                            .filter_map(|(f, r)| f(&p).then_some(*r))
-                            .fold(BitFlags::<Reason>::default(), |a, b| a | b);
-                        let empty = reason.is_empty();
-                        if !empty {
-                            Some((Identifier::Transcript(tx.clone()), reason))
-                        } else {
-                            None
-                        }
-                    })
-                    .collect_vec();
-                if tx_reasons.is_empty() {
-                    None
-                } else {
-                    Some(tx_reasons)
-                }
-            })
-            .flatten()
-            .collect_vec();
-
-        for (id, reason) in discarded {
-            self.mark_discarded(&id, reason)?;
-        }
-        // self.discard()?;
-
-        tracing::info!("… done filtering transcripts in {:?}", start.elapsed());
-        Ok(())
-    }
-
-    fn filter_transcripts_with_sequence(
-        &mut self,
-        seq_provider: &mut SequenceProvider,
-    ) -> Result<HashMap<TranscriptId, String>, Error> {
-        tracing::info!("Filtering transcripts with sequences …");
-        let start = Instant::now();
-
-        let five_prime_truncated = |tx: &Transcript| -> bool {
-            let is_mt = MITOCHONDRIAL_ACCESSIONS.iter().any(|a| tx.is_on_contig(a));
-            if tx.protein_coding() && !is_mt {
-                tx.genome_builds.iter().any(|(_release, alignment)| {
-                    let cds_start = alignment.cds_start;
-                    let mut exons = alignment.exons.clone();
-                    exons.is_empty() || {
-                        exons.sort_unstable_by_key(|e| e.alt_start_i);
-                        let first = exons.first().unwrap();
-                        Some(first.alt_start_i) == cds_start
-                    }
-                })
-            } else {
-                false
-            }
-        };
-        let three_prime_truncated = |tx: &Transcript| -> bool {
-            let is_mt = MITOCHONDRIAL_ACCESSIONS.iter().any(|a| tx.is_on_contig(a));
-            if tx.protein_coding() && !is_mt {
-                tx.genome_builds.iter().any(|(_release, alignment)| {
-                    let cds_end = alignment.cds_end;
-                    let mut exons = alignment.exons.clone();
-                    exons.is_empty() || {
-                        exons.sort_unstable_by_key(|e| e.alt_end_i);
-                        let last = exons.last().unwrap();
-                        Some(last.alt_end_i) == cds_end
-                    }
-                })
-            } else {
-                false
-            }
-        };
-
-        // Check transcript's CDS length for being a multiple of 3.
-        //
-        // Note that the chrMT transcripts have been fixed earlier already to
-        // accommodate for how they are fixed by poly-A tailing.
-        let invalid_cds_length = |tx: &Transcript| -> bool {
-            if tx.protein_coding() {
-                tx.cds_length().is_none_or(|l| l % 3 != 0)
-            } else {
-                false
-            }
-        };
-
-        let (discards, keeps_with_reasons): (Vec<_>, Vec<_>) = self
-            .transcript_id_to_transcript
-            .par_iter()
-            .partition_map(|(tx_id, tx)| {
-                if let Some(d) = self.discards.get(&Identifier::Transcript(tx_id.clone()))
-                    && d.intersects(Reason::hard())
-                {
-                    return Either::Left((Identifier::Transcript(tx_id.clone()), *d));
-                }
-
-                let has_invalid_cds_length = if self.disable_filters {
-                    false
-                } else {
-                    invalid_cds_length(tx)
-                };
-
-                let namespace: String = if tx_id.starts_with("ENST") {
-                    String::from("Ensembl")
-                } else {
-                    String::from("NCBI")
-                };
-                let seq = seq_provider.fetch_sequence(&AliasOrSeqId::Alias {
-                    value: (*tx_id).to_string(),
-                    namespace: Some(namespace.clone()),
-                });
-
-                if let Ok(seq) = seq {
-                    let is_mt = MITOCHONDRIAL_ACCESSIONS.iter().any(|a| tx.is_on_contig(a));
-                    let is_seleno = tx
-                        .biotype
-                        .as_ref()
-                        .map(|bt| bt.contains(&BioType::Selenoprotein))
-                        .unwrap_or(false);
-                    if seq.is_empty() {
-                        return Either::Left((
-                            Identifier::Transcript(tx_id.clone()),
-                            Reason::MissingSequence.into(),
-                        ));
-                    };
-                    // Skip transcript if it is coding and the translated CDS does not have a stop codon.
-                    let cds = tx.start_codon.and_then(|start| {
-                        tx.stop_codon.map(|stop| (start as usize, stop as usize))
-                    });
-                    let cds = if tx.protein_coding() { cds } else { None };
-                    if let Some((cds_start, cds_end)) = cds {
-                        let cds_length = cds_end.saturating_sub(cds_start);
-                        let delta = (3 - (cds_length % 3)).max(cds_end.saturating_sub(seq.len()));
-                        let seq = append_poly_a(seq, delta);
-
-                        let safe_end = cds_end.min(seq.len());
-                        let safe_start = cds_start.min(safe_end);
-
-                        let tx_seq_to_translate = &seq[safe_start..safe_end];
-
-                        let aa_sequence_result = translate_cds(
-                            tx_seq_to_translate,
-                            true,
-                            "*",
-                            if is_mt {
-                                TranslationTable::VertebrateMitochondrial
-                            } else if is_seleno {
-                                TranslationTable::Selenocysteine
-                            } else {
-                                TranslationTable::Standard
-                            },
-                        );
-
-                        let mut reason = Reason::empty();
-
-                        if let Ok(aa_sequence) = aa_sequence_result {
-                            let has_missing_stop_codon = (!is_mt && !aa_sequence.ends_with('*'))
-                                || (is_mt && !aa_sequence.contains('*'));
-                            if has_missing_stop_codon && !self.disable_filters {
-                                reason |= Reason::MissingStopCodon;
-                                if five_prime_truncated(tx) {
-                                    reason |= Reason::FivePrimeEndTruncated;
-                                }
-                                if three_prime_truncated(tx) {
-                                    reason |= Reason::ThreePrimeEndTruncated;
-                                }
-                            }
-                        }
-
-                        if has_invalid_cds_length
-                            || self
-                                .fixes
-                                .get(&Identifier::Transcript(tx_id.clone()))
-                                .is_some_and(|f| f.contains(Fix::Cds))
-                        {
-                            reason |= Reason::InvalidCdsLength;
-                        }
-
-                        if reason.intersects(Reason::hard()) {
-                            Either::Left((Identifier::Transcript(tx_id.clone()), reason))
-                        } else {
-                            Either::Right((tx_id.clone(), seq, reason))
-                        }
-                    } else {
-                        Either::Right((tx_id.clone(), seq, Reason::empty()))
-                    }
-                } else {
-                    Either::Left((
-                        Identifier::Transcript(tx_id.clone()),
-                        Reason::MissingSequence.into(),
-                    ))
-                }
-            });
-
-        for (id, reason) in discards {
-            self.mark_discarded(&id, reason)?;
-        }
-        // self.discard()?;
-
-        let mut sequence_map = HashMap::new();
-        for (tx_id, seq, reason) in keeps_with_reasons {
-            if !reason.is_empty() {
-                self.mark_discarded(&Identifier::Transcript(tx_id.clone()), reason)?;
-            }
-            sequence_map.insert(tx_id, seq);
-        }
-
-        tracing::info!(
-            "… done filtering transcripts with sequence provider in {:?}",
-            start.elapsed()
-        );
-
-        Ok(sequence_map)
-    }
-
-    /// Remove all gene ids for which there are no transcripts left.
-    fn filter_empty_gene_id_mappings(&mut self) -> Result<(), Error> {
-        tracing::info!("Removing empty Gene Id mappings …");
-        let gene_ids_gene: HashSet<_> = self.gene_id_to_gene.keys().collect();
-        let gene_ids_tx: HashSet<_> = self.gene_id_to_transcript_ids.keys().collect();
-        let gene_id_but_no_tx_id = &gene_ids_gene - &gene_ids_tx;
-        let tx_id_but_no_gene_id = &gene_ids_tx - &gene_ids_gene;
-        for gene_id in gene_id_but_no_tx_id {
-            *self
-                .discards
-                .entry(Identifier::Gene(gene_id.clone()))
-                .or_default() |= Reason::NoTranscripts;
-        }
-        for gene_id in tx_id_but_no_gene_id {
-            *self
-                .discards
-                .entry(Identifier::Gene(gene_id.clone()))
-                .or_default() |= Reason::MissingGene;
-        }
-
-        for (gene_id, txs) in self.gene_id_to_transcript_ids.iter() {
-            if txs.is_empty() {
-                *self
-                    .discards
-                    .entry(Identifier::Gene(gene_id.clone()))
-                    .or_default() |= Reason::NoTranscriptLeft;
-            }
-        }
-        // self.discard()?;
-        tracing::info!("… done removing empty HGNC mappings");
-        Ok(())
     }
 
     fn update_pseudogene_status(&mut self) -> Result<(), Error> {
@@ -900,28 +389,6 @@ impl TranscriptLoader {
         }
     }
 
-    /// Group transcripts by release and version (sorted descending).
-    fn by_release_and_version<'a>(
-        txs: &[&'a Transcript],
-    ) -> HashMap<(String, String), Vec<(u32, &'a Transcript)>> {
-        let mut versioned = txs
-            .iter()
-            .flat_map(|tx| {
-                let releases = tx.genome_builds.keys().cloned().collect_vec();
-                let tx_id = TranscriptId::try_new(&tx.id).unwrap();
-                let (ac, version) = tx_id.split_version();
-                let ac = ac.to_string();
-                releases
-                    .into_iter()
-                    .map(move |r| ((r, ac.clone()), (version, *tx)))
-            })
-            .into_group_map();
-        versioned.values_mut().for_each(|txs| {
-            txs.sort_unstable_by_key(|(version, _)| Reverse(*version));
-        });
-        versioned
-    }
-
     fn discard(&mut self, remove: bool) -> Result<(), Error> {
         let (n_transcripts_pre, n_gene_ids_pre) = (
             self.transcript_id_to_transcript.len(),
@@ -1088,341 +555,28 @@ impl TranscriptLoader {
 
         Ok(())
     }
+}
 
-    /// Perform protobuf file construction.
-    ///
-    /// This can be done by simply converting the models from ``hgvs-rs`` to the prost generated data structures.
-    fn build_protobuf(
-        &mut self,
-        sequence_map: &mut HashMap<TranscriptId, String>,
-        source_version: SourceVersion,
-    ) -> Result<TxSeqDatabase, Error> {
-        tracing::info!("Constructing protobuf data structures …");
-        let start = Instant::now();
-
-        let (gene_ids, tx_ids): (Vec<GeneId>, Vec<TranscriptId>) = {
-            // ensure we use all gene_ids that are available to us;
-            // this includes gene_ids for which we only have transcripts but no genes or vice versa
-            let gene_ids = self
-                .gene_id_to_gene
-                .keys()
-                .cloned()
-                .chain(self.gene_id_to_transcript_ids.keys().cloned())
-                .sorted_unstable()
-                .dedup()
-                .collect_vec();
-
-            // We do, however, discard transcripts that have no sequence or gene id.
-            let tx_ids = self
-                .gene_id_to_transcript_ids
-                .values()
-                .flatten()
-                .sorted_unstable()
-                .dedup()
-                .filter(|&tx_id| {
-                    self.discards
-                        .get(&Identifier::Transcript(tx_id.clone()))
-                        .is_none_or(|reason| !reason.intersects(Reason::hard()))
-                })
-                .cloned()
-                .collect();
-            (gene_ids, tx_ids)
-        };
-
-        // Construct sequence database.
-        tracing::info!("  Constructing sequence database …");
-        let seq_db = {
-            // filter out transcripts that have no sequence, enumerate, collect; build SequenceDb
-            let (aliases, aliases_idx, seqs) = tx_ids
-                .iter()
-                .filter_map(|tx_id| {
-                    sequence_map
-                        .remove(tx_id)
-                        .map(|seq| ((*tx_id).to_string(), seq))
-                })
-                .enumerate()
-                .map(|(idx, (tx_id, seq))| (tx_id, idx as u32, seq))
-                .multiunzip();
-
-            crate::pbs::txs::SequenceDb {
-                aliases,
-                aliases_idx,
-                seqs,
-            }
-        };
-
-        tracing::info!("  Creating transcript records for each gene…");
-        let empty_txs = vec![];
-        let data_transcripts = gene_ids
-            .iter()
-            .flat_map(|gene_id| {
-                let tx_ids = self
-                    .gene_id_to_transcript_ids
-                    .get(gene_id)
-                    .unwrap_or(&empty_txs);
-                tx_ids
-                    .iter()
-                    .sorted_unstable()
-                    .map(|tx_id| self.protobuf_transcript(gene_id, tx_id))
-            })
-            .collect();
-
-        tracing::info!(" … done creating transcripts in {:#?}", start.elapsed());
-
-        // Build mapping of gene HGNC symbol to transcript IDs.
-        tracing::info!("  Build gene symbol to transcript ID mapping …");
-        let empty = vec![];
-        let gene_to_tx = gene_ids
-            .iter()
-            .map(|gene_id| {
-                let tx_ids = self
-                    .gene_id_to_transcript_ids
-                    .get(gene_id)
-                    .unwrap_or(&empty);
-                let gene_id_reason = self.discards.get(&Identifier::Gene(gene_id.clone()));
-                let filtered =
-                    gene_id_reason.is_some_and(|reason| reason.intersects(Reason::hard()));
-                crate::pbs::txs::GeneToTxId {
-                    gene_id: gene_id.to_string(),
-                    tx_ids: tx_ids
-                        .iter()
-                        .sorted_unstable()
-                        .map(|tx_id| tx_id.to_string())
-                        .collect(),
-                    filtered: Some(filtered),
-                    filter_reason: gene_id_reason.map(|r| r.bits()),
-                }
-            })
-            .collect::<Vec<_>>();
-        tracing::info!(" … done building gene symbol to transcript ID mapping");
-
-        // Compose transcript database from transcripts and gene to transcript mapping.
-        tracing::info!("  Composing transcript seq database …");
-        let tx_db = crate::pbs::txs::TranscriptDb {
-            transcripts: data_transcripts,
-            gene_to_tx,
-        };
-
-        // Compose the final transcript and sequence database.
-        let tx_seq_db = TxSeqDatabase {
-            tx_db: Some(tx_db),
-            seq_db: Some(seq_db),
-            version: Some(crate::common::version().to_string()),
-            source_version: vec![source_version],
-        };
-
-        tracing::info!(" … done composing transcript seq database");
-
-        Ok(tx_seq_db)
-    }
-
-    fn protobuf_transcript(
-        &self,
-        gene_id: &GeneId,
-        tx_id: &TranscriptId,
-    ) -> crate::pbs::txs::Transcript {
-        let tx = self
-            .transcript_id_to_transcript
-            .get(tx_id)
-            .unwrap_or_else(|| panic!("No transcript for id {:?}", tx_id));
-
-        // Combine discard reasons for transcript and gene level.
-        let tx_reason = self
-            .discards
-            .get(&Identifier::Transcript(tx_id.clone()))
-            .copied()
-            .unwrap_or_default();
-        let gene_reason = self
-            .discards
-            .get(&Identifier::Gene(gene_id.clone()))
-            .copied()
-            .unwrap_or_default();
-        let combined_reason = tx_reason | gene_reason;
-
-        // Only filter out the transcript if it intersects with a "hard" reason.
-        // Soft-filtered transcripts will have `filtered = false`, but their `filter_reason`
-        // will still contain the bitmask for downstream inspection.
-        let filtered = combined_reason.intersects(Reason::hard());
-
-        // ... build genome alignment for selected:
-        let (genome_alignments, tags): (Vec<_>, Vec<_>) = tx
-            .genome_builds
-            .iter()
-            .map(|(genome_build, alignment)| {
-                Self::protobuf_genome_alignment(genome_build, alignment)
-            })
-            .unzip();
-
-        let seleno_tag = tx
-            .biotype
-            .as_ref()
-            .is_some_and(|bt| bt.contains(&BioType::Selenoprotein))
-            .then(|| crate::pbs::txs::TranscriptTag::Selenoprotein.into());
-
-        let tags = tags
-            .into_iter()
-            .flatten()
-            .chain(seleno_tag)
-            .sorted_unstable()
-            .dedup()
-            .collect();
-
-        // Now, just obtain the basic properties and create a new `data::Transcript`.
-        let Gene { gene_symbol, .. } =
-            self.gene_id_to_gene
-                .get(gene_id)
-                .cloned()
-                .unwrap_or_else(|| Gene {
-                    aliases: None,
-                    biotype: None,
-                    description: None,
-                    gene_symbol: None,
-                    hgnc: Some(gene_id.to_string()),
-                    map_location: None,
-                    summary: None,
-                    url: "".to_string(),
-                });
-
-        let Transcript {
-            protein,
-            start_codon,
-            stop_codon,
-            ..
-        } = tx.clone();
-        let biotype = if tx.protein_coding() {
-            crate::pbs::txs::TranscriptBiotype::Coding.into()
-        } else {
-            crate::pbs::txs::TranscriptBiotype::NonCoding.into()
-        };
-
-        crate::pbs::txs::Transcript {
-            id: (*tx_id).to_string(),
-            gene_symbol: gene_symbol.unwrap_or("<MISSING>".to_string()),
-            gene_id: gene_id.to_string(),
-            biotype,
-            tags,
-            protein,
-            start_codon,
-            stop_codon,
-            genome_alignments,
-            filtered: Some(filtered),
-            filter_reason: combined_reason
-                .is_empty()
-                .not()
-                .then(|| combined_reason.bits()),
-        }
-    }
-
-    fn protobuf_genome_alignment(
-        genome_build: &str,
-        alignment: &GenomeAlignment,
-    ) -> (crate::pbs::txs::GenomeAlignment, HashSet<i32>) {
-        let mut tags = HashSet::new();
-        let genome_build = genome_build.to_lowercase();
-
-        let GenomeAlignment {
-            contig,
-            cds_start,
-            cds_end,
-            ..
-        } = alignment.clone();
-        let strand = match alignment.strand {
-            cdot_models::Strand::Plus => crate::pbs::txs::Strand::Plus,
-            cdot_models::Strand::Minus => crate::pbs::txs::Strand::Minus,
-        };
-        if let Some(tag) = alignment.tag.as_ref() {
-            for t in tag {
-                let elem = match t {
-                    Tag::Basic => crate::pbs::txs::TranscriptTag::Basic.into(),
-                    Tag::EnsemblCanonical => {
-                        crate::pbs::txs::TranscriptTag::EnsemblCanonical.into()
-                    }
-                    Tag::ManeSelect => crate::pbs::txs::TranscriptTag::ManeSelect.into(),
-                    Tag::ManePlusClinical => {
-                        crate::pbs::txs::TranscriptTag::ManePlusClinical.into()
-                    }
-                    Tag::RefSeqSelect => crate::pbs::txs::TranscriptTag::RefSeqSelect.into(),
-                    Tag::GencodePrimary => crate::pbs::txs::TranscriptTag::GencodePrimary.into(),
-                    Tag::Other(v) => match v.as_str() {
-                        "EnsemblGraft" => crate::pbs::txs::TranscriptTag::EnsemblGraft.into(),
-                        "basic-backport" => crate::pbs::txs::TranscriptTag::BasicBackport.into(),
-                        "ensembl_canonical-backport" => {
-                            crate::pbs::txs::TranscriptTag::EnsemblCanonicalBackport.into()
-                        }
-                        "mane_select-backport" => {
-                            crate::pbs::txs::TranscriptTag::ManeSelectBackport.into()
-                        }
-                        "mane_plus_clinical-backport" => {
-                            crate::pbs::txs::TranscriptTag::ManePlusClinicalBackport.into()
-                        }
-                        "ref_seq_select-backport" => {
-                            crate::pbs::txs::TranscriptTag::RefSeqSelectBackport.into()
-                        }
-                        "selenoprotein-backport" => {
-                            crate::pbs::txs::TranscriptTag::SelenoproteinBackport.into()
-                        }
-                        "gencode_primary-backport" => {
-                            crate::pbs::txs::TranscriptTag::GencodePrimaryBackport.into()
-                        }
-                        x if x.ends_with("-backport") => {
-                            crate::pbs::txs::TranscriptTag::OtherBackport.into()
-                        }
-                        _ => crate::pbs::txs::TranscriptTag::Other.into(),
-                    },
-                };
-                tags.insert(elem);
-            }
-        }
-        // Look into any "note" string for a selenoprotein marker and
-        // add this as a tag.
-        if let Some(note) = alignment.note.as_ref() {
-            let needle = "UGA stop codon recoded as selenocysteine";
-            if note.contains(needle) {
-                tags.insert(crate::pbs::txs::TranscriptTag::Selenoprotein.into());
-            }
-        }
-        // and construct vector of all exons
-        let exons: Vec<_> = alignment
-            .exons
-            .iter()
-            .map(|exon| {
-                let cdot_models::Exon {
-                    alt_start_i,
-                    alt_end_i,
-                    ord,
-                    alt_cds_start_i,
-                    alt_cds_end_i,
-                    cigar,
-                } = exon.clone();
-                crate::pbs::txs::ExonAlignment {
-                    alt_start_i,
-                    alt_end_i,
-                    ord,
-                    alt_cds_start_i: if alt_cds_start_i == -1 {
-                        None
-                    } else {
-                        Some(alt_cds_start_i)
-                    },
-                    alt_cds_end_i: if alt_cds_end_i == -1 {
-                        None
-                    } else {
-                        Some(alt_cds_end_i)
-                    },
-                    cigar,
-                }
-            })
-            .collect();
-
-        let genome_alignment = crate::pbs::txs::GenomeAlignment {
-            genome_build,
-            contig,
-            cds_start,
-            cds_end,
-            strand: strand.into(),
-            exons,
-        };
-        (genome_alignment, tags)
-    }
+/// Group transcripts by release and version (sorted descending).
+fn group_transcripts_by_release_and_version<'a>(
+    txs: &[&'a Transcript],
+) -> HashMap<(String, String), Vec<(u32, &'a Transcript)>> {
+    let mut versioned = txs
+        .iter()
+        .flat_map(|tx| {
+            let releases = tx.genome_builds.keys().cloned().collect_vec();
+            let tx_id = TranscriptId::try_new(&tx.id).unwrap();
+            let (ac, version) = tx_id.split_version();
+            let ac = ac.to_string();
+            releases
+                .into_iter()
+                .map(move |r| ((r, ac.clone()), (version, *tx)))
+        })
+        .into_group_map();
+    versioned.values_mut().for_each(|txs| {
+        txs.sort_unstable_by_key(|(version, _)| Reverse(*version));
+    });
+    versioned
 }
 
 fn append_poly_a(seq: String, length: usize) -> String {
@@ -1565,42 +719,44 @@ pub fn run(common: &crate::common::Args, args: &Args) -> Result<(), Error> {
             "total_gene_ids": tx_data.gene_id_to_transcript_ids.len()
         })))?;
 
+        let tx_data_ = &mut tx_data;
+
         // … then filter gene id entries with no transcripts to boot …
-        tx_data.filter_initial_gene_id_entries()?;
+        filter_initial_gene_id_entries(tx_data_)?;
         // … then filter genes (missing gene id and/or symbol) …
-        tx_data.filter_genes()?;
+        filter_genes(tx_data_)?;
         // … then filter transcripts …
-        tx_data.filter_transcripts()?;
+        filter_transcripts(tx_data_)?;
         // … ensure there are no gene keys without associated transcripts left …
-        tx_data.filter_empty_gene_id_mappings()?;
+        filter_empty_gene_id_mappings(tx_data_)?;
 
         // Open seqrepo / FASTA …
         let mut seq_provider = reference::open_sequence_provider(args)?;
         // … and filter transcripts based on their sequences,
         // e.g. checking whether their translation contains a stop codon …
-        let mut sequence_map = tx_data.filter_transcripts_with_sequence(&mut seq_provider)?;
-        tx_data.filter_empty_gene_id_mappings()?;
+        let mut sequence_map = filter_transcripts_with_sequence(tx_data_, &mut seq_provider)?;
+        filter_empty_gene_id_mappings(tx_data_)?;
         // … if there are genes with no transcripts left, check whether they are pseudogenes …
-        tx_data.update_pseudogene_status()?;
+        tx_data_.update_pseudogene_status()?;
 
         // … trigger the discard routine, but do not remove anything, just make sure they are consistent,
         // and report the stats.
         let remove = false;
-        tx_data.discard(remove)?;
+        tx_data_.discard(remove)?;
 
         // … and update all discard annotations …
-        tx_data.propagate_discard_reasons(&raw_tx_data)?;
+        tx_data_.propagate_discard_reasons(&raw_tx_data)?;
 
         trace_rss_now();
 
         report(ReportEntry::Log(json!({
             "source": "annotations_filtered",
-            "total_transcripts": tx_data.transcript_id_to_transcript.len(),
-            "total_gene_ids": tx_data.gene_id_to_transcript_ids.len()
+            "total_transcripts": tx_data_.transcript_id_to_transcript.len(),
+            "total_gene_ids": tx_data_.gene_id_to_transcript_ids.len()
         })))?;
 
         // For some stats, count number of chrMt, MANE Select and MANE Plus Clinical transcripts.
-        let (n_mt, n_mane_select, n_mane_plus_clinical) = tx_data.gather_transcript_stats()?;
+        let (n_mt, n_mane_select, n_mane_plus_clinical) = tx_data_.gather_transcript_stats()?;
 
         report(ReportEntry::Log(json!({
             "source": "annotations_filtered",
@@ -1612,7 +768,7 @@ pub fn run(common: &crate::common::Args, args: &Args) -> Result<(), Error> {
         let source_version = bundle_source_version_information(args);
 
         // … and finally construct protobuf txdb data structures.
-        let tx_db = tx_data.build_protobuf(&mut sequence_map, source_version)?;
+        let tx_db = build_protobuf(tx_data_, &mut sequence_map, source_version)?;
 
         // List all discarded transcripts and genes.
         for (id, reason) in tx_data.discards.into_iter().sorted_unstable() {
@@ -1727,6 +883,7 @@ pub mod test {
     use crate::db::create::TranscriptLoader;
     use crate::db::create::cdot::load_cdot;
     use crate::db::create::cli::Args;
+    use crate::db::create::filter::filter_transcripts;
     use crate::db::create::models::GeneId;
     use crate::db::dump;
 
@@ -1755,7 +912,7 @@ pub mod test {
                 .collect::<Vec<_>>()
         );
 
-        tx_data.filter_transcripts()?;
+        filter_transcripts(&mut tx_data)?;
         insta::assert_yaml_snapshot!(
             tx_data
                 .gene_id_to_transcript_ids
